Use -localapi_socket arg for the local API socket path
There was some severely confusing thing going on with -localapi_socket
before, where it would default to the public API hostname, but then if
you passed a different socket pathname as a cmdline parameter, it
wouldn't connect to that.

This change straightens out the strings (defaulting to the reasonable
default path for the local API socket, passing a value to the UNIX
domain socket dialer if passed).